# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conduit Management API",
    "version": "0.1.0",
    "description": "CRUD for tenants, API keys, MCP servers, rate limits, and audit log queries. Runs on its own port so it can be isolated from agent traffic."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api-keys": {
      "get": { "summary": "List API keys for a tenant", "parameters": [{ "name": "tenant_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "OK" } } },
      "post": { "summary": "Create an API key (raw key returned once)", "responses": { "201": { "description": "Created" } } }
    },
    "/api-keys/{keyID}": {
      "delete": { "summary": "Revoke an API key", "responses": { "204": { "description": "No Content" }, "404": { "description": "Not found" } } }
    },
...
    "/rate-limits": {
      "get": { "summary": "List rate limit configs for a tenant", "parameters": [{ "name": "tenant_id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "summary": "Create or update a rate limit config", "responses": { "200": { "description": "OK" } } }
    },
    "/rate-limits/{id}": {
      "delete": { "summary": "Delete a rate limit config", "responses": { "204": { "description": "No Content" }, "404": { "description": "Not found" } } }
    },
...
  }
}